Compare Data By Fetching From Exel And Db
In this project, it is a .NET application that detects the difference between the data corresponding to the relevant id in Excel and the data corresponding to the relevant id in the db and outputs the total number of different data and the relevant id value. With this project, 14798 data numbers can be manipulated in 2 seconds.